fix forward declarations
for an aggregate we want to skip forward declarations.  they will
always cause problems with D and there is no reason we should want
them.

what is a forward declaration?  it's a second declaration of an
aggregate where there is a first declaration (earlier or later in
the translation unit) that defines it.  in that case we should
definitely skip the forward declaration for an aggregate.

for a function - which may be a member function aka method, we
only want the forward declarations.  I don't see how it can cause
any harm for a function (the prototype has to be the same), but
currently free-standing methods with the implementation somewhere
else end up in the file, only without even a this pointer.

this change prevents that happenining.

neeeds tests.