Where aggregate methods in C++ are defined outside the definition of the aggregate they can end up being defined the same way in the .d file, which is illegal.
This commit skips the cursor when there is both a declaration and a
definition of the function in the TU, and the cursor is the declaration.

Conceptually cursor.definition and Cursor.nullcursor are weakly pure,
but they are not declared to be so in the wrapper/binding.  So I made
a little callAssumePure helper which will call an impure delegate
assuming it is actually weakly pure.

private bool isRedundantFunctionDeclaration(const(from!"clang".Cursor)
cursor) @trusted
{
    import clang: Cursor;
    // skip cursor when there is both a declaration and a definition and
the cursor is the declaration
    return (cursor.definition != Cursor.nullCursor) && (cursor ==
cursor.definition);
}