This commit adds gathering sorted by namespace (global namespace comes last because it tends to clutter readability of the output).
It also adds version using new extern(C++,...) syntax.  This has
the consequence of no longer mapping C++ namespaces to D module
namespace which means that a struct Bar in namespace Foo can no
longer be referred to as a Foo.Bar within the main .d file.

I thought that it actually does make sense to map C++ namespaces
to D modules and have multiple .d files generated, one per namespace.
But for the time being this new feature is versioned out so will
not compile.