Modifications to make `dub test` pass on Windows
Note it will also require new unit-threaded and libclang
to fully pass.

In a couple cases, I exempted the test since it was not
applicable on the Windows platform, but for the most part,
it was tweaking platform-specific details like object file
name extensions and hacking around a slower filesystem.

The MS headers use literals in the form of, e.g. 255i8, and
that extension needed to be handled as well. Otherwise, fairly
small overall.